requirements.txt
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gcloud==0.17.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
passlib==1.7.4
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dotenv==0.15.0
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
